requirements.txt
future
emoji
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
pynewtonmath
#memes vvv
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
#translator
googletrans
jikanpy
speedtest-cli
tswift